+++ ot/__init__.py
@@ -35,6 +35,7 @@ from . import weak
 from . import factored
 from . import solvers
 from . import gaussian
+from . import lowrank
 
 
 # OT functions
@@ -51,7 +52,7 @@ from .gromov import (gromov_wasserstein, gromov_wasserstein2,
                      gromov_barycenters, fused_gromov_wasserstein, fused_gromov_wasserstein2)
 from .weak import weak_optimal_transport
 from .factored import factored_optimal_transport
-from .solvers import solve, solve_gromov
+from .solvers import solve, solve_gromov, solve_sample
 from .lowrank import lowrank_sinkhorn
 
 # utils functions